Extracted reused functions to new helper class
